huggingface_space/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
gradio>=4.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.11.0
timm>=0.9.0
